requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-import-export==2.6.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
